Update folder structure and improve your code
